This patch addresses two issues:
A small socket receive buffer can result in a buffer overflow (and lost ping replies) if many fast targets are pinged.  The receive buffer has accordingly been enlarged to 131072 bytes.  In the long run this should probably be varied according to the number of pinged targets.

The second issue is that the errno values are currently hard-coded.  On platforms other than Windows and Linux this causes trouble.  Rather use the values defined in the errno package: errno.EWOULDBLOCK, which resolves to 11 on Linux, 10035 on Windows, 35 on FreeBSD, and other values on other platforms.