Error removed
"no responses pending" should not be handled as an error. If someone uses multi_ping with retry != 0, he/she does it to make sure that a possibly sent ICMP packages does not get lost. But when the first trial worked, the function should only exit and not throw by any means an error.