refs #23: fixed multi_ping retry
Now the multi_ping retry functionality works correctly.
The issue was that an ICMP retry packet was sent with a new ID, but the
receive function was still checking packets using old IDs.

Now the remainings_id is reinitialized with the new ids during each sends.
This way the receive function always used the latest ids for packet
matching.

The source address of the packet has also been added to the packet
matching to make the request/reply match more robust.